fix(sdk): move inline-edit-controller into the SDK, not extension-host
Phase 2 (255f7ee) put this in extension-host, but the dependency only
runs one direction — extension-host depends on @silo-code/sdk, never
the reverse — and InlineEdit (Phase 3, a public SDK component) is the
thing that will call setActiveInlineEditCancel. Relocating now so
Phase 3 starts from a correct base instead of hitting this as a
package-boundary violation mid-implementation.

Exported from the SDK barrel tagged @internal: real exported surface
(so host code can import it like anything else in @silo-code/sdk), but
excluded from the docs reference — extension authors never call these
directly. Modal.tsx now imports yieldEscapeToInlineEdit from
"@silo-code/sdk" instead of a local host file.
